refactor(migrate): finish upsampler modularization
Split the upsampler facade from position rewriting.

Move node mapping and unit tests into focused modules.

Mark csl26-rfct upsampler complete in the same change.
