+++ crates/citum-migrate/src/upsampler.rs
@@ -3,20 +3,23 @@ SPDX-License-Identifier: MIT OR Apache-2.0
 SPDX-FileCopyrightText: © 2023-2026 Bruce D'Arcus and Citum contributors
 */
 
-use citum_schema::{self as csln, FormattingOptions, ItemType, Variable};
-use csl_legacy::model::{self as legacy, CslNode as LNode};
-use std::collections::HashMap;
+use citum_schema as citum;
+use csl_legacy::model::CslNode as LNode;
 use std::sync::OnceLock;
 
+mod mapping;
 mod position;
 
-use position::{
-    CitationPositionTarget, analyze_citation_positions, branch_is_effectively_unconditional,
-    choose_has_position_condition, choose_uses_position_fast_path, is_supported_position_token,
-    select_position_branch_children,
-};
+#[cfg(test)]
+mod tests;
 
-fn migrate_debug_enabled() -> bool {
+pub use position::CitationPositionTemplates;
+
+/// Citum-facing alias for schema nodes emitted by the upsampler.
+pub type CitumNode = citum::CslnNode;
+
+/// Returns true when verbose migration debug logging is enabled.
+pub(super) fn migrate_debug_enabled() -> bool {
     static ENABLED: OnceLock<bool> = OnceLock::new();
     *ENABLED.get_or_init(|| {
         std::env::var("CITUM_MIGRATE_DEBUG")
@@ -36,43 +39,6 @@ pub struct Upsampler {
     provenance: Option<crate::ProvenanceTracker>,
     pub et_al_min: Option<usize>,
     pub et_al_use_first: Option<usize>,
-}
-
-/// Citation templates extracted from CSL `position` conditions.
-///
-/// These templates are used to populate Citum citation position overrides:
-/// `first`/default branch -> base citation template,
-/// `subsequent` branch -> `citation.subsequent`,
-/// `ibid`/`ibid-with-locator` branches -> `citation.ibid`.
-#[derive(Debug, Clone, Default)]
-pub struct CitationPositionTemplates {
-    /// Template nodes for the first/default citation form.
-    pub first: Option<Vec<csln::CslnNode>>,
-    /// Template nodes for subsequent non-immediate repeats.
-    pub subsequent: Option<Vec<csln::CslnNode>>,
-    /// Template nodes for immediate repeats (`ibid`, `ibid-with-locator`).
-    pub ibid: Option<Vec<csln::CslnNode>>,
-    /// Whether the source tree mixed `position` with unsupported conditions.
-    pub unsupported_mixed_conditions: bool,
-}
-
-fn unsupported_position_templates() -> CitationPositionTemplates {
-    CitationPositionTemplates {
-        unsupported_mixed_conditions: true,
-        ..Default::default()
-    }
-}
-
-fn unsupported_position_result<T>(result: Result<T, ()>) -> Result<T, CitationPositionTemplates> {
-    result.map_err(|()| unsupported_position_templates())
-}
-
-impl CitationPositionTemplates {
-    /// Returns true when at least one position-specific override is available.
-    #[must_use]
-    pub fn has_overrides(&self) -> bool {
-        self.subsequent.is_some() || self.ibid.is_some()
-    }
 }
 
 impl Upsampler {
@@ -98,8 +64,8 @@ impl Upsampler {
 
     /// The entry point for converting a flattened legacy tree into Citum nodes.
     #[must_use]
-    pub fn upsample_nodes(&self, legacy_nodes: &[LNode]) -> Vec<csln::CslnNode> {
-        let mut csln_nodes = Vec::new();
+    pub fn upsample_nodes(&self, legacy_nodes: &[LNode]) -> Vec<CitumNode> {
+        let mut citum_nodes = Vec::new();
         let mut i = 0;
 
         while i < legacy_nodes.len() {
@@ -109,1647 +75,18 @@ impl Upsampler {
             if let LNode::Group(group) = node
                 && let Some(collapsed) = self.try_collapse_label_variable(group)
             {
-                csln_nodes.push(collapsed);
+                citum_nodes.push(collapsed);
                 i += 1;
                 continue;
             }
 
             if let Some(mapped) = self.map_node(node) {
-                csln_nodes.push(mapped);
+                citum_nodes.push(mapped);
             }
 
             i += 1;
         }
 
-        csln_nodes
-    }
-
-    /// Extract position-scoped citation templates from legacy CSL `choose` trees.
-    ///
-    /// Pure position-only trees still use direct branch selection. Mixed trees
-    /// are specialized by stripping `position` from matching branches while
-    /// preserving their remaining predicates and non-position sibling content.
-    #[must_use]
-    pub fn extract_citation_position_templates(
-        &self,
-        legacy_nodes: &[LNode],
-    ) -> CitationPositionTemplates {
-        let analysis = analyze_citation_positions(legacy_nodes);
-
-        if !analysis.has_position_chooses() {
-            return CitationPositionTemplates::default();
-        }
-
-        if analysis.has_unsupported_mixed_conditions() {
-            return unsupported_position_templates();
-        }
-
-        let Ok(first) =
-            self.extract_position_variant_if(legacy_nodes, true, CitationPositionTarget::First)
-        else {
-            return unsupported_position_templates();
-        };
-
-        let Ok(subsequent) = self.extract_position_variant_if(
-            legacy_nodes,
-            analysis.has_explicit_subsequent(),
-            CitationPositionTarget::Subsequent,
-        ) else {
-            return unsupported_position_templates();
-        };
-
-        let Ok(ibid) = self.extract_position_variant_if(
-            legacy_nodes,
-            analysis.has_explicit_ibid(),
-            CitationPositionTarget::IbidAny,
-        ) else {
-            return unsupported_position_templates();
-        };
-
-        CitationPositionTemplates {
-            first,
-            subsequent,
-            ibid,
-            unsupported_mixed_conditions: false,
-        }
-    }
-
-    fn upsample_position_variant(
-        &self,
-        legacy_nodes: &[LNode],
-        target: CitationPositionTarget,
-    ) -> Result<Option<Vec<csln::CslnNode>>, ()> {
-        let rewritten = self.rewrite_nodes_for_position(legacy_nodes, target)?;
-        let upsampled = self.upsample_nodes(&rewritten);
-        if upsampled.is_empty() {
-            Ok(None)
-        } else {
-            Ok(Some(upsampled))
-        }
-    }
-
-    fn extract_position_variant_if(
-        &self,
-        legacy_nodes: &[LNode],
-        should_extract: bool,
-        target: CitationPositionTarget,
-    ) -> Result<Option<Vec<csln::CslnNode>>, CitationPositionTemplates> {
-        if !should_extract {
-            return Ok(None);
-        }
-
-        unsupported_position_result(self.upsample_position_variant(legacy_nodes, target))
-    }
-
-    fn rewrite_nodes_for_position(
-        &self,
-        legacy_nodes: &[LNode],
-        target: CitationPositionTarget,
-    ) -> Result<Vec<LNode>, ()> {
-        let mut rewritten = Vec::new();
-
-        for node in legacy_nodes {
-            if let Some(rewritten_container) = self.rewrite_child_container(node, target)? {
-                rewritten.push(rewritten_container);
-                continue;
-            }
-
-            match node {
-                LNode::Choose(choose) if choose_has_position_condition(choose) => {
-                    if choose_uses_position_fast_path(choose) {
-                        let selected = select_position_branch_children(choose, target);
-                        rewritten.extend(self.rewrite_nodes_for_position(selected, target)?);
-                    } else {
-                        rewritten.extend(self.rewrite_mixed_position_choose(choose, target)?);
-                    }
-                }
-                LNode::Choose(choose) => {
-                    rewritten.push(self.rewrite_non_position_choose(choose, target)?);
-                }
-                _ => rewritten.push(node.clone()),
-            }
-        }
-
-        Ok(rewritten)
-    }
-
-    fn rewrite_child_container(
-        &self,
-        node: &LNode,
-        target: CitationPositionTarget,
-    ) -> Result<Option<LNode>, ()> {
-        let rewritten = match node {
-            LNode::Group(group) => {
-                let mut rewritten_group = group.clone();
-                rewritten_group.children =
-                    self.rewrite_nodes_for_position(&group.children, target)?;
-                Some(LNode::Group(rewritten_group))
-            }
-            LNode::Names(names) => {
-                let mut rewritten_names = names.clone();
-                rewritten_names.children =
-                    self.rewrite_nodes_for_position(&names.children, target)?;
-                Some(LNode::Names(rewritten_names))
-            }
-            LNode::Substitute(substitute) => {
-                let mut rewritten_substitute = substitute.clone();
-                rewritten_substitute.children =
-                    self.rewrite_nodes_for_position(&substitute.children, target)?;
-                Some(LNode::Substitute(rewritten_substitute))
-            }
-            _ => None,
-        };
-
-        Ok(rewritten)
-    }
-
-    fn rewrite_choose_branch_children(
-        &self,
-        branch: &legacy::ChooseBranch,
-        target: CitationPositionTarget,
-    ) -> Result<legacy::ChooseBranch, ()> {
-        let mut rewritten_branch = branch.clone();
-        rewritten_branch.children = self.rewrite_nodes_for_position(&branch.children, target)?;
-        Ok(rewritten_branch)
-    }
-
-    fn rewrite_non_position_choose(
-        &self,
-        choose: &legacy::Choose,
-        target: CitationPositionTarget,
-    ) -> Result<LNode, ()> {
-        let mut rewritten_choose = choose.clone();
-        rewritten_choose.if_branch =
-            self.rewrite_choose_branch_children(&choose.if_branch, target)?;
-        rewritten_choose.else_if_branches = choose
-            .else_if_branches
-            .iter()
-            .map(|branch| self.rewrite_choose_branch_children(branch, target))
-            .collect::<Result<Vec<_>, _>>()?;
-        rewritten_choose.else_branch = choose
-            .else_branch
-            .as_ref()
-            .map(|branch| self.rewrite_nodes_for_position(branch, target))
-            .transpose()?;
-        Ok(LNode::Choose(rewritten_choose))
-    }
-
-    fn rewrite_mixed_position_branch(
-        &self,
-        branch: &legacy::ChooseBranch,
-        target: CitationPositionTarget,
-    ) -> Result<Option<legacy::ChooseBranch>, ()> {
-        let Some(position) = &branch.position else {
-            return self
-                .rewrite_choose_branch_children(branch, target)
-                .map(Some);
-        };
-
-        let matched_target =
-            position
-                .split_whitespace()
-                .try_fold(false, |matched_target, token| {
-                    if !is_supported_position_token(token) {
-                        return Err(());
-                    }
-                    Ok(matched_target || target.matches_token(token))
-                })?;
-
-        if !matched_target {
-            return Ok(None);
-        }
-
-        let mut rewritten_branch = self.rewrite_choose_branch_children(branch, target)?;
-        rewritten_branch.position = None;
-        Ok(Some(rewritten_branch))
-    }
-
-    fn assemble_rewritten_position_choose(
-        &self,
-        mut rewritten_branches: Vec<legacy::ChooseBranch>,
-        rewritten_else: Option<Vec<LNode>>,
-    ) -> Result<Vec<LNode>, ()> {
-        let unconditional_positions = rewritten_branches
-            .iter()
-            .enumerate()
-            .filter_map(|(index, branch)| {
-                branch_is_effectively_unconditional(branch).then_some(index)
-            })
-            .collect::<Vec<_>>();
-
-        if unconditional_positions.len() + usize::from(rewritten_else.is_some()) > 1 {
-            return Err(());
-        }
-
-        if rewritten_branches.is_empty() {
-            return Ok(rewritten_else.unwrap_or_default());
-        }
-
-        if let Some(index) = unconditional_positions.first().copied() {
-            if index == 0 {
-                return Ok(rewritten_branches.remove(0).children);
-            }
-
-            let else_branch = rewritten_branches.remove(index).children;
-            let mut branches = rewritten_branches.into_iter();
-            let Some(if_branch) = branches.next() else {
-                return Ok(else_branch);
-            };
-
-            return Ok(vec![LNode::Choose(legacy::Choose {
-                if_branch,
-                else_if_branches: branches.collect(),
-                else_branch: Some(else_branch),
-            })]);
-        }
-
-        let mut branches = rewritten_branches.into_iter();
-        let Some(if_branch) = branches.next() else {
-            return Ok(rewritten_else.unwrap_or_default());
-        };
-
-        Ok(vec![LNode::Choose(legacy::Choose {
-            if_branch,
-            else_if_branches: branches.collect(),
-            else_branch: rewritten_else,
-        })])
-    }
-
-    fn rewrite_mixed_position_choose(
-        &self,
-        choose: &legacy::Choose,
-        target: CitationPositionTarget,
-    ) -> Result<Vec<LNode>, ()> {
-        let mut rewritten_branches = Vec::new();
-
-        for branch in std::iter::once(&choose.if_branch).chain(choose.else_if_branches.iter()) {
-            if let Some(rewritten_branch) = self.rewrite_mixed_position_branch(branch, target)? {
-                rewritten_branches.push(rewritten_branch);
-            }
-        }
-
-        let rewritten_else = choose
-            .else_branch
-            .as_ref()
-            .map(|branch| self.rewrite_nodes_for_position(branch, target))
-            .transpose()?;
-        self.assemble_rewritten_position_choose(rewritten_branches, rewritten_else)
-    }
-
-    fn map_node(&self, node: &LNode) -> Option<csln::CslnNode> {
-        match node {
-            LNode::Text(t) => {
-                if let Some(var_str) = &t.variable
-                    && let Some(var) = self.map_variable(var_str)
-                {
-                    if let Some(ref prov) = self.provenance {
-                        let var_name = format!("{var:?}").to_lowercase();
-                        prov.record_upsampling(&var_name, "Text", "Variable");
-                    }
-                    if migrate_debug_enabled() {
-                        tracing::debug!(
-                            "Upsampler: Text({:?}) macro_call_order={:?}",
-                            var,
-                            t.macro_call_order
-                        );
-                    }
-                    return Some(csln::CslnNode::Variable(csln::VariableBlock {
-                        variable: var,
-                        label: None,
-                        formatting: self.map_formatting(
-                            &t.formatting,
-                            &t.prefix,
-                            &t.suffix,
-                            t.quotes,
-                            t.strip_periods,
-                        ),
-                        overrides: HashMap::new(),
-                        source_order: t.macro_call_order,
-                    }));
-                }
-                if let Some(term) = &t.term {
-                    if let Some(general_term) = csln::locale::Locale::parse_general_term(term) {
-                        return Some(csln::CslnNode::Term(csln::TermBlock {
-                            term: general_term,
-                            form: self.map_term_form(t.form.as_deref()),
-                            formatting: self.map_formatting(
-                                &t.formatting,
-                                &t.prefix,
-                                &t.suffix,
-                                t.quotes,
-                                t.strip_periods,
-                            ),
-                            source_order: t.macro_call_order,
-                        }));
-                    }
-
-                    // Fallback for unknown terms
-                    let prefix = t.prefix.as_deref().unwrap_or("");
-                    let suffix = t.suffix.as_deref().unwrap_or("");
-                    let text_cased = self.apply_text_case(term, t.text_case.as_deref());
-                    return Some(csln::CslnNode::Text {
-                        value: format!("{prefix}{text_cased}{suffix}"),
-                    });
-                }
-                if let Some(val) = &t.value {
-                    return Some(csln::CslnNode::Text { value: val.clone() });
-                }
-                None
-            }
-            LNode::Group(g) => Some(csln::CslnNode::Group(csln::GroupBlock {
-                children: self.upsample_nodes(&g.children),
-                delimiter: g.delimiter.clone(),
-                formatting: self.map_formatting(&g.formatting, &g.prefix, &g.suffix, None, None),
-                source_order: g.macro_call_order,
-            })),
-            LNode::Date(d) => self.map_date(d),
-            LNode::Names(n) => self.map_names(n),
-            LNode::Choose(c) => self.map_choose(c),
-            LNode::Number(n) => self.map_number(n),
-            LNode::Label(l) => self.map_label(l),
-            _ => None,
-        }
-    }
-
-    fn apply_name_child_options(
-        &self,
-        n: &legacy::Names,
-        variable: &csln::Variable,
-        options: &mut csln::NamesOptions,
-        et_al_min: &mut Option<usize>,
-        et_al_use_first: &mut Option<usize>,
-        et_al_term: &mut String,
-    ) {
-        for child in &n.children {
-            match child {
-                LNode::Name(name) => {
-                    options.mode = match name.form.as_deref() {
-                        Some("short") => Some(csln::NameMode::Short),
-                        Some("count") => Some(csln::NameMode::Count),
-                        _ => Some(csln::NameMode::Long),
-                    };
-                    options.and = match name.and.as_deref() {
-                        Some("text") => Some(csln::AndTerm::Text),
-                        Some("symbol") => Some(csln::AndTerm::Symbol),
-                        _ => None,
-                    };
-                    options.initialize_with = name.initialize_with.clone();
-                    options.sort_separator = name.sort_separator.clone();
-                    options.name_as_sort_order = match name.name_as_sort_order.as_deref() {
-                        Some("first") => Some(csln::NameAsSortOrder::First),
-                        Some("all") => Some(csln::NameAsSortOrder::All),
-                        _ => None,
-                    };
-                    options.delimiter_precedes_last = match name.delimiter_precedes_last.as_deref()
-                    {
-                        Some("contextual") => Some(csln::DelimiterPrecedes::Contextual),
-                        Some("after-inverted-name") => {
-                            Some(csln::DelimiterPrecedes::AfterInvertedName)
-                        }
-                        Some("always") => Some(csln::DelimiterPrecedes::Always),
-                        Some("never") => Some(csln::DelimiterPrecedes::Never),
-                        _ => None,
-                    };
-
-                    // Name node can also have et-al attributes
-                    if name.et_al_min.is_some() {
-                        *et_al_min = name.et_al_min;
-                    }
-                    if name.et_al_use_first.is_some() {
-                        *et_al_use_first = name.et_al_use_first;
-                    }
-                }
-                LNode::Label(label) => {
-                    options.label = Some(csln::LabelOptions {
-                        variable: variable.clone(),
-                        form: self.map_label_form(&label.form),
-                        pluralize: true,
-                        formatting: self.map_formatting(
-                            &label.formatting,
-                            &label.prefix,
-                            &label.suffix,
-                            None,
-                            label.strip_periods,
-                        ),
-                    });
-                }
-                LNode::EtAl(et_al) => {
-                    if let Some(term) = &et_al.term {
-                        *et_al_term = term.clone();
-                    }
-                    // Formatting from et-al node? Legacy model needs to capture it.
-                    // For now, default.
-                }
-                LNode::Substitute(sub) => {
-                    for sub_node in &sub.children {
-                        if let LNode::Names(sub_names) = sub_node
-                            && let Some(sub_var) = self.map_variable(&sub_names.variable)
-                        {
-                            options.substitute.push(sub_var);
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn map_names(&self, n: &legacy::Names) -> Option<csln::CslnNode> {
-        let vars: Vec<&str> = n.variable.split_whitespace().collect();
-        if vars.is_empty() {
-            return None;
-        }
-
-        #[allow(clippy::indexing_slicing, reason = "vars is not empty")]
-        let variable = self.map_variable(vars[0])?;
-
-        let mut options = csln::NamesOptions {
-            delimiter: n.delimiter.clone(),
-            ..Default::default()
-        };
-
-        // If multiple variables were provided, add the others to substitute
-        for v in vars.iter().skip(1) {
-            if let Some(var) = self.map_variable(v) {
-                options.substitute.push(var);
-            }
-        }
-
-        // Extract et-al defaults from Names node, falling back to upsampler defaults
-        let mut et_al_min = n.et_al_min.or(self.et_al_min);
-        let mut et_al_use_first = n.et_al_use_first.or(self.et_al_use_first);
-        let et_al_subsequent =
-            if n.et_al_subsequent_min.is_some() || n.et_al_subsequent_use_first.is_some() {
-                let fallback_min = et_al_min.unwrap_or(0) as u8;
-                let fallback_use_first = et_al_use_first.unwrap_or(0) as u8;
-                Some(Box::new(csln::EtAlSubsequent {
-                    min: n.et_al_subsequent_min.map_or(fallback_min, |v| v as u8),
-                    use_first: n
-                        .et_al_subsequent_use_first
-                        .map_or(fallback_use_first, |v| v as u8),
-                }))
-            } else {
-                None
-            };
-
-        let mut et_al_term = "et al.".to_string();
-        let et_al_formatting = FormattingOptions::default();
-
-        self.apply_name_child_options(
-            n,
-            &variable,
-            &mut options,
-            &mut et_al_min,
-            &mut et_al_use_first,
-            &mut et_al_term,
-        );
-
-        if let Some(min) = et_al_min {
-            options.et_al = Some(csln::EtAlOptions {
-                min: min as u8,
-                use_first: et_al_use_first.unwrap_or(1) as u8,
-                subsequent: et_al_subsequent,
-                term: et_al_term,
-                formatting: et_al_formatting,
-            });
-        }
-
-        if migrate_debug_enabled() {
-            tracing::debug!(
-                "Upsampler: Names({:?}) macro_call_order={:?}",
-                variable,
-                n.macro_call_order
-            );
-        }
-        Some(csln::CslnNode::Names(csln::NamesBlock {
-            variable,
-            options,
-            formatting: FormattingOptions::default(),
-            source_order: n.macro_call_order,
-        }))
-    }
-
-    fn map_number(&self, n: &legacy::Number) -> Option<csln::CslnNode> {
-        let variable = self.map_variable(&n.variable)?;
-        Some(csln::CslnNode::Variable(csln::VariableBlock {
-            variable,
-            label: None,
-            formatting: self.map_formatting(&n.formatting, &n.prefix, &n.suffix, None, None),
-            overrides: HashMap::new(),
-            source_order: n.macro_call_order,
-        }))
-    }
-
-    fn map_label(&self, l: &legacy::Label) -> Option<csln::CslnNode> {
-        if let Some(var_str) = &l.variable
-            && let Some(var) = self.map_variable(var_str)
-        {
-            return Some(csln::CslnNode::Variable(csln::VariableBlock {
-                variable: var.clone(),
-                label: Some(csln::LabelOptions {
-                    variable: var,
-                    form: self.map_label_form(&l.form),
-                    pluralize: true,
-                    formatting: self.map_formatting(
-                        &l.formatting,
-                        &l.prefix,
-                        &l.suffix,
-                        None,
-                        l.strip_periods,
-                    ),
-                }),
-                formatting: FormattingOptions::default(),
-                overrides: HashMap::new(),
-                source_order: l.macro_call_order,
-            }));
-        }
-        None
-    }
-
-    fn upsample_first_node(&self, nodes: &[LNode]) -> Option<csln::CslnNode> {
-        self.upsample_nodes(nodes).into_iter().next()
-    }
-
-    fn map_uncertain_date_choose(&self, choose: &legacy::Choose) -> Option<csln::CslnNode> {
-        choose.if_branch.is_uncertain_date.as_ref()?;
-
-        choose
-            .else_branch
-            .as_deref()
-            .and_then(|else_children| self.upsample_first_node(else_children))
-            .or_else(|| {
-                choose
-                    .else_if_branches
-                    .first()
-                    .and_then(|branch| self.upsample_first_node(&branch.children))
-            })
-            .or_else(|| {
-                // A bare uncertain-date branch means "emit this only for uncertain dates".
-                // Migration defaults to the common certain-date case, so the absence of an
-                // else branch should compile to no output rather than unconditional output.
-                Some(csln::CslnNode::Group(csln::GroupBlock {
-                    children: vec![],
-                    delimiter: None,
-                    formatting: FormattingOptions::default(),
-                    source_order: None,
-                }))
-            })
-    }
-
-    fn map_position_choose_fallback(&self, choose: &legacy::Choose) -> Option<csln::CslnNode> {
-        if !choose_has_position_condition(choose) {
-            return None;
-        }
-
-        choose
-            .else_branch
-            .as_deref()
-            .and_then(|else_children| self.upsample_first_node(else_children))
-            .or_else(|| {
-                choose
-                    .else_if_branches
-                    .iter()
-                    .find(|branch| branch.position.is_none())
-                    .and_then(|branch| self.upsample_first_node(&branch.children))
-            })
-    }
-
-    fn map_branch_item_types(&self, type_names: Option<&str>) -> Vec<ItemType> {
-        type_names
-            .into_iter()
-            .flat_map(|types| types.split_whitespace())
-            .filter_map(|item_type| self.map_item_type(item_type))
-            .collect()
-    }
-
-    fn map_branch_variables(&self, variable_names: Option<&str>) -> Vec<Variable> {
-        variable_names
-            .into_iter()
-            .flat_map(|variables| variables.split_whitespace())
-            .filter_map(|variable| self.map_variable(variable))
-            .collect()
-    }
-
-    fn map_negated_else_if_fallback(
-        &self,
-        branch: &legacy::ChooseBranch,
-    ) -> Option<Vec<csln::CslnNode>> {
-        (branch.match_mode.as_deref() == Some("none") && branch.type_.is_some())
-            .then(|| self.upsample_nodes(&branch.children))
-    }
-
-    fn map_else_if_branch(&self, branch: &legacy::ChooseBranch) -> csln::ElseIfBranch {
-        csln::ElseIfBranch {
-            if_item_type: self.map_branch_item_types(branch.type_.as_deref()),
-            if_variables: self.map_branch_variables(branch.variable.as_deref()),
-            children: self.upsample_nodes(&branch.children),
-        }
-    }
-
-    fn resolve_condition_else_branch(
-        &self,
-        choose: &legacy::Choose,
-        negated_else_nodes: Option<Vec<csln::CslnNode>>,
-    ) -> Option<Vec<csln::CslnNode>> {
-        choose
-            .else_branch
-            .as_ref()
-            .map(|branch| self.upsample_nodes(branch))
-            .or(negated_else_nodes)
-    }
-
-    fn resolve_condition_branches(
-        &self,
-        choose: &legacy::Choose,
-        if_match_none: bool,
-        else_branch: Option<Vec<csln::CslnNode>>,
-    ) -> (Vec<csln::CslnNode>, Option<Vec<csln::CslnNode>>) {
-        if if_match_none {
-            let if_nodes = self.upsample_nodes(&choose.if_branch.children);
-            (Vec::new(), else_branch.or(Some(if_nodes)))
-        } else {
-            (self.upsample_nodes(&choose.if_branch.children), else_branch)
-        }
-    }
-
-    fn map_choose(&self, c: &legacy::Choose) -> Option<csln::CslnNode> {
-        // Handle is-uncertain-date condition specially: prefer else branch since most dates
-        // aren't uncertain. Full EDTF support would handle this dynamically at render time.
-        if let Some(node) = self.map_uncertain_date_choose(c) {
-            return Some(node);
-        }
-
-        // Handle position conditions (ibid, subsequent, etc.) by preferring else branch.
-        // Position conditions are for repeated citations - else branch has full first-citation.
-        if let Some(node) = self.map_position_choose_fallback(c) {
-            return Some(node);
-        }
-
-        // Determine if the if-branch uses match="none" (negated type test).
-        // A negated if-branch fires for everything NOT in its type list, so it
-        // behaves like a default/else branch rather than a type-specific branch.
-        let if_match_none = c.if_branch.match_mode.as_deref() == Some("none");
-
-        let if_item_type = if if_match_none {
-            Vec::new()
-        } else {
-            self.map_branch_item_types(c.if_branch.type_.as_deref())
-        };
-        let if_variables = self.map_branch_variables(c.if_branch.variable.as_deref());
-
-        // Map all else-if branches. For branches with match="none" (negated type
-        // condition), clear the type list — they act as broad defaults, not as
-        // type-specific branches. This ensures compile_for_type selects them as
-        // the else/fallback path for types not covered by positive branches.
-        let mut else_if_branches: Vec<csln::ElseIfBranch> = Vec::new();
-        let mut negated_else_nodes: Option<Vec<csln::CslnNode>> = None;
-
-        for branch in &c.else_if_branches {
-            if let Some(fallback_nodes) = self.map_negated_else_if_fallback(branch) {
-                // Treat this as a fallback else branch, since it fires for all
-                // types NOT in its type list (i.e., the "default" case).
-                // Only adopt the first such branch to avoid duplicates.
-                if negated_else_nodes.is_none() {
-                    negated_else_nodes = Some(fallback_nodes);
-                }
-                continue;
-            }
-
-            else_if_branches.push(self.map_else_if_branch(branch));
-        }
-
-        // Determine the effective else_branch: prefer the existing else branch,
-        // then fall back to the negated else-if content if present.
-        let else_branch = self.resolve_condition_else_branch(c, negated_else_nodes);
-
-        // Handle the if-branch match="none" case: push the if-branch content as
-        // the else fallback, since it fires for all non-listed types.
-        let (then_branch, else_branch) =
-            self.resolve_condition_branches(c, if_match_none, else_branch);
-
-        Some(csln::CslnNode::Condition(csln::ConditionBlock {
-            if_item_type,
-            if_variables,
-            then_branch,
-            else_if_branches,
-            else_branch,
-        }))
-    }
-
-    fn map_item_type(&self, s: &str) -> Option<ItemType> {
-        match s {
-            "article" => Some(ItemType::Article),
-            "article-journal" => Some(ItemType::ArticleJournal),
-            "article-magazine" => Some(ItemType::ArticleMagazine),
-            "article-newspaper" => Some(ItemType::ArticleNewspaper),
-            "bill" => Some(ItemType::Bill),
-            "book" => Some(ItemType::Book),
-            "broadcast" => Some(ItemType::Broadcast),
-            "chapter" => Some(ItemType::Chapter),
-            "dataset" => Some(ItemType::Dataset),
-            "entry" => Some(ItemType::Entry),
-            "entry-dictionary" => Some(ItemType::EntryDictionary),
-            "entry-encyclopedia" => Some(ItemType::EntryEncyclopedia),
-            "figure" => Some(ItemType::Figure),
-            "graphic" => Some(ItemType::Graphic),
-            "interview" => Some(ItemType::Interview),
-            "legal_case" => Some(ItemType::LegalCase),
-            "legislation" => Some(ItemType::Legislation),
-            "manuscript" => Some(ItemType::Manuscript),
-            "map" => Some(ItemType::Map),
-            "motion_picture" => Some(ItemType::MotionPicture),
-            "musical_score" => Some(ItemType::MusicalScore),
-            "pamphlet" => Some(ItemType::Pamphlet),
-            "paper-conference" => Some(ItemType::PaperConference),
-            "patent" => Some(ItemType::Patent),
-            "personal_communication" => Some(ItemType::PersonalCommunication),
-            "post" => Some(ItemType::Post),
-            "post-weblog" => Some(ItemType::PostWeblog),
-            "report" => Some(ItemType::Report),
-            "review" => Some(ItemType::Review),
-            "review-book" => Some(ItemType::ReviewBook),
-            "song" => Some(ItemType::Song),
-            "software" => Some(ItemType::Software),
-            "speech" => Some(ItemType::Speech),
-            "standard" => Some(ItemType::Standard),
-            "thesis" => Some(ItemType::Thesis),
-            "treaty" => Some(ItemType::Treaty),
-            "webpage" => Some(ItemType::Webpage),
-            _ => None,
-        }
-    }
-
-    fn map_date(&self, d: &legacy::Date) -> Option<csln::CslnNode> {
-        let variable = self.map_variable(&d.variable)?;
-        let mut year_form = None;
-        let mut month_form = None;
-        let mut day_form = None;
-
-        for part in &d.parts {
-            match part.name.as_str() {
-                "year" => year_form = self.map_date_part_form(&part.form),
-                "month" => month_form = self.map_date_part_form(&part.form),
-                "day" => day_form = self.map_date_part_form(&part.form),
-                _ => {}
-            }
-        }
-
-        if migrate_debug_enabled() {
-            tracing::debug!(
-                "Upsampler: Date({:?}) macro_call_order={:?}",
-                variable,
-                d.macro_call_order
-            );
-        }
-        Some(csln::CslnNode::Date(csln::DateBlock {
-            variable,
-            options: csln::DateOptions {
-                form: match d.form.as_deref() {
-                    Some("text") => Some(csln::DateForm::Text),
-                    Some("numeric") => Some(csln::DateForm::Numeric),
-                    _ => None,
-                },
-                parts: match d.date_parts.as_deref() {
-                    Some("year") => Some(csln::DateParts::Year),
-                    Some("year-month") => Some(csln::DateParts::YearMonth),
-                    _ => None,
-                },
-                delimiter: d.delimiter.clone(),
-                year_form,
-                month_form,
-                day_form,
-            },
-            formatting: self.map_formatting(&d.formatting, &d.prefix, &d.suffix, None, None),
-            source_order: d.macro_call_order,
-        }))
-    }
-
-    fn map_date_part_form(&self, form: &Option<String>) -> Option<csln::DatePartForm> {
-        match form.as_deref() {
-            Some("numeric") => Some(csln::DatePartForm::Numeric),
-            Some("numeric-leading-zeros") => Some(csln::DatePartForm::NumericLeadingZeros),
-            Some("ordinal") => Some(csln::DatePartForm::Ordinal),
-            Some("long") => Some(csln::DatePartForm::Long),
-            Some("short") => Some(csln::DatePartForm::Short),
-            _ => None,
-        }
-    }
-
-    fn try_collapse_label_variable(&self, group: &legacy::Group) -> Option<csln::CslnNode> {
-        if group.children.len() == 2 {
-            #[allow(clippy::indexing_slicing, reason = "group.children.len() == 2")]
-            let first = &group.children[0];
-            #[allow(clippy::indexing_slicing, reason = "group.children.len() == 2")]
-            let second = &group.children[1];
-
-            if let (LNode::Label(l), LNode::Text(t)) = (first, second)
-                && let (Some(l_var), Some(t_var)) = (&l.variable, &t.variable)
-                && l_var == t_var
-                && let Some(var) = self.map_variable(t_var)
-            {
-                return Some(csln::CslnNode::Variable(csln::VariableBlock {
-                    variable: var.clone(),
-                    label: Some(csln::LabelOptions {
-                        variable: var,
-                        form: self.map_label_form(&l.form),
-                        pluralize: true,
-                        formatting: self.map_formatting(
-                            &l.formatting,
-                            &l.prefix,
-                            &l.suffix,
-                            None,
-                            l.strip_periods,
-                        ),
-                    }),
-                    formatting: self.map_formatting(
-                        &t.formatting,
-                        &t.prefix,
-                        &t.suffix,
-                        t.quotes,
-                        t.strip_periods,
-                    ),
-                    overrides: HashMap::new(),
-                    source_order: t.macro_call_order,
-                }));
-            }
-        }
-        None
-    }
-
-    fn map_variable(&self, s: &str) -> Option<Variable> {
-        match s {
-            "title" => Some(Variable::Title),
-            "container-title" => Some(Variable::ContainerTitle),
-            "collection-title" => Some(Variable::CollectionTitle),
-            "original-title" => Some(Variable::OriginalTitle),
-            "publisher" => Some(Variable::Publisher),
-            "publisher-place" => Some(Variable::PublisherPlace),
-            "archive" => Some(Variable::Archive),
-            "archive-place" => Some(Variable::ArchivePlace),
-            "archive_location" => Some(Variable::ArchiveLocation),
-            "event" => Some(Variable::Event),
-            "event-place" => Some(Variable::EventPlace),
-            "page" => Some(Variable::Page),
-            "locator" => Some(Variable::Locator),
-            "version" => Some(Variable::Version),
-            "volume" => Some(Variable::Volume),
-            "number-of-volumes" => Some(Variable::NumberOfVolumes),
-            "issue" => Some(Variable::Issue),
-            "chapter-number" => Some(Variable::ChapterNumber),
-            "medium" => Some(Variable::Medium),
-            "status" => Some(Variable::Status),
-            "edition" => Some(Variable::Edition),
-            "section" => Some(Variable::Section),
-            "source" => Some(Variable::Source),
-            "genre" => Some(Variable::Genre),
-            "note" => Some(Variable::Note),
-            "annote" => Some(Variable::Annote),
-            "abstract" => Some(Variable::Abstract),
-            "keyword" => Some(Variable::Keyword),
-            "number" => Some(Variable::Number),
-            "URL" => Some(Variable::URL),
-            "DOI" => Some(Variable::DOI),
-            "ISBN" => Some(Variable::ISBN),
-            "ISSN" => Some(Variable::ISSN),
-            "PMID" => Some(Variable::PMID),
-            "PMCID" => Some(Variable::PMCID),
-            "call-number" => Some(Variable::CallNumber),
-            "dimensions" => Some(Variable::Dimensions),
-            "scale" => Some(Variable::Scale),
-            "jurisdiction" => Some(Variable::Jurisdiction),
-            "citation-label" => Some(Variable::CitationLabel),
-            "citation-number" => Some(Variable::CitationNumber),
-            "year-suffix" => Some(Variable::YearSuffix),
-            "author" => Some(Variable::Author),
-            "editor" => Some(Variable::Editor),
-            "editorial-director" => Some(Variable::EditorialDirector),
-            "translator" => Some(Variable::Translator),
-            "illustrator" => Some(Variable::Illustrator),
-            "original-author" => Some(Variable::OriginalAuthor),
-            "container-author" => Some(Variable::ContainerAuthor),
-            "collection-editor" => Some(Variable::CollectionEditor),
-            "composer" => Some(Variable::Composer),
-            "director" => Some(Variable::Director),
-            "interviewer" => Some(Variable::Interviewer),
-            "recipient" => Some(Variable::Recipient),
-            "reviewed-author" => Some(Variable::ReviewedAuthor),
-            "issued" => Some(Variable::Issued),
-            "event-date" => Some(Variable::EventDate),
-            "accessed" => Some(Variable::Accessed),
-            "container" => Some(Variable::Submitted),
-            "original-date" => Some(Variable::OriginalDate),
-            "available-date" => Some(Variable::AvailableDate),
-            _ => None,
-        }
-    }
-
-    fn map_label_form(&self, form: &Option<String>) -> csln::LabelForm {
-        match form.as_deref() {
-            Some("short") => csln::LabelForm::Short,
-            Some("symbol") => csln::LabelForm::Symbol,
-            _ => csln::LabelForm::Long,
-        }
-    }
-
-    /// Apply text-case transformation to a string.
-    /// Handles CSL 1.0 text-case attribute values for term nodes.
-    fn apply_text_case(&self, text: &str, case: Option<&str>) -> String {
-        match case {
-            Some("capitalize-first") => {
-                let mut chars = text.chars();
-                match chars.next() {
-                    None => String::new(),
-                    Some(first) => first.to_uppercase().collect::<String>() + chars.as_str(),
-                }
-            }
-            Some("capitalize-all") => text
-                .split_whitespace()
-                .map(|word| {
-                    let mut chars = word.chars();
-                    match chars.next() {
-                        None => String::new(),
-                        Some(first) => first.to_uppercase().collect::<String>() + chars.as_str(),
-                    }
-                })
-                .collect::<Vec<_>>()
-                .join(" "),
-            Some("lowercase") => text.to_lowercase(),
-            Some("uppercase") => text.to_uppercase(),
-            _ => text.to_string(),
-        }
-    }
-
-    fn map_formatting(
-        &self,
-        f: &legacy::Formatting,
-        prefix: &Option<String>,
-        suffix: &Option<String>,
-        quotes: Option<bool>,
-        strip_periods: Option<bool>,
-    ) -> FormattingOptions {
-        FormattingOptions {
-            font_style: f.font_style.as_ref().map(|s| match s.as_str() {
-                "italic" => csln::FontStyle::Italic,
-                "oblique" => csln::FontStyle::Oblique,
-                _ => csln::FontStyle::Normal,
-            }),
-            font_weight: f.font_weight.as_ref().map(|s| match s.as_str() {
-                "bold" => csln::FontWeight::Bold,
-                "light" => csln::FontWeight::Light,
-                _ => csln::FontWeight::Normal,
-            }),
-            font_variant: f.font_variant.as_ref().map(|s| match s.as_str() {
-                "small-caps" => csln::FontVariant::SmallCaps,
-                _ => csln::FontVariant::Normal,
-            }),
-            text_decoration: f.text_decoration.as_ref().map(|s| match s.as_str() {
-                "underline" => csln::TextDecoration::Underline,
-                _ => csln::TextDecoration::None,
-            }),
-            vertical_align: f.vertical_align.as_ref().map(|s| match s.as_str() {
-                "superscript" => csln::VerticalAlign::Superscript,
-                "subscript" => csln::VerticalAlign::Subscript,
-                _ => csln::VerticalAlign::Baseline,
-            }),
-            quotes,
-            prefix: prefix.clone(),
-            suffix: suffix.clone(),
-            strip_periods,
-        }
-    }
-    fn map_term_form(&self, form: Option<&str>) -> csln::locale::TermForm {
-        match form {
-            Some("short") => csln::locale::TermForm::Short,
-            Some("verb") => csln::locale::TermForm::Verb,
-            Some("verb-short") => csln::locale::TermForm::VerbShort,
-            Some("symbol") => csln::locale::TermForm::Symbol,
-            _ => csln::locale::TermForm::Long,
-        }
-    }
-}
-
-#[cfg(test)]
-#[allow(
-    clippy::unwrap_used,
-    clippy::expect_used,
-    clippy::panic,
-    clippy::indexing_slicing,
-    clippy::todo,
-    clippy::unimplemented,
-    clippy::unreachable,
-    clippy::get_unwrap,
-    reason = "Panicking is acceptable and often desired in tests."
-)]
-mod tests {
-    use super::*;
-    use csl_legacy::model::{Choose, ChooseBranch, CslNode, Formatting, Group, Text};
-
-    fn literal_text(value: &str) -> CslNode {
-        CslNode::Text(Text {
-            value: Some(value.to_string()),
-            variable: None,
-            macro_name: None,
-            term: None,
-            form: None,
-            prefix: None,
-            suffix: None,
-            quotes: None,
-            text_case: None,
-            strip_periods: None,
-            plural: None,
-            macro_call_order: None,
-            formatting: Formatting::default(),
-        })
-    }
-
-    fn choose_branch(position: Option<&str>, children: Vec<CslNode>) -> ChooseBranch {
-        ChooseBranch {
-            match_mode: None,
-            type_: None,
-            variable: None,
-            is_numeric: None,
-            is_uncertain_date: None,
-            locator: None,
-            position: position.map(ToString::to_string),
-            children,
-        }
-    }
-
-    fn choose_branch_with_conditions(
-        position: Option<&str>,
-        type_: Option<&str>,
-        variable: Option<&str>,
-        locator: Option<&str>,
-        children: Vec<CslNode>,
-    ) -> ChooseBranch {
-        ChooseBranch {
-            match_mode: None,
-            type_: type_.map(ToString::to_string),
-            variable: variable.map(ToString::to_string),
-            is_numeric: None,
-            is_uncertain_date: None,
-            locator: locator.map(ToString::to_string),
-            position: position.map(ToString::to_string),
-            children,
-        }
-    }
-
-    fn group(children: Vec<CslNode>) -> CslNode {
-        CslNode::Group(Group {
-            delimiter: None,
-            prefix: None,
-            suffix: None,
-            children,
-            macro_call_order: None,
-            formatting: Formatting::default(),
-        })
-    }
-
-    fn collect_text_values<'a>(nodes: &'a [csln::CslnNode], output: &mut Vec<&'a str>) {
-        for node in nodes {
-            match node {
-                csln::CslnNode::Text { value } => output.push(value.as_str()),
-                csln::CslnNode::Group(group) => collect_text_values(&group.children, output),
-                csln::CslnNode::Condition(condition) => {
-                    collect_text_values(&condition.then_branch, output);
-                    for branch in &condition.else_if_branches {
-                        collect_text_values(&branch.children, output);
-                    }
-                    if let Some(else_branch) = &condition.else_branch {
-                        collect_text_values(else_branch, output);
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn text_values(nodes: &[csln::CslnNode]) -> Vec<&str> {
-        let mut values = Vec::new();
-        collect_text_values(nodes, &mut values);
-        values
-    }
-
-    #[test]
-    fn extract_position_templates_preserves_nested_choose_siblings() {
-        let choose = Choose {
-            if_branch: choose_branch(Some("subsequent"), vec![literal_text("SHORT")]),
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-        let legacy_nodes = vec![group(vec![
-            literal_text("PREFIX"),
-            CslNode::Choose(choose),
-            literal_text("SUFFIX"),
-        ])];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert_eq!(
-            text_values(
-                extracted
-                    .first
-                    .as_deref()
-                    .expect("first variant should exist")
-            ),
-            vec!["PREFIX", "FULL", "SUFFIX"]
-        );
-        assert_eq!(
-            text_values(
-                extracted
-                    .subsequent
-                    .as_deref()
-                    .expect("subsequent variant should exist")
-            ),
-            vec!["PREFIX", "SHORT", "SUFFIX"]
-        );
-    }
-
-    #[test]
-    fn extract_position_templates_merges_multiple_position_chooses() {
-        let choose = Choose {
-            if_branch: choose_branch(Some("subsequent"), vec![literal_text("SHORT")]),
-            else_if_branches: vec![choose_branch(
-                Some("first"),
-                vec![literal_text("FIRST-ONLY")],
-            )],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-        let locator_choose = Choose {
-            if_branch: choose_branch(Some("ibid-with-locator"), vec![literal_text("LOC")]),
-            else_if_branches: vec![choose_branch(Some("ibid"), vec![literal_text("IBID")])],
-            else_branch: Some(vec![literal_text("DATE")]),
-        };
-        let legacy_nodes = vec![
-            literal_text("A"),
-            CslNode::Choose(choose),
-            literal_text("MID"),
-            CslNode::Choose(locator_choose),
-            literal_text("Z"),
-        ];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert_eq!(
-            text_values(
-                extracted
-                    .first
-                    .as_deref()
-                    .expect("first variant should exist")
-            ),
-            vec!["A", "FIRST-ONLY", "MID", "DATE", "Z"],
-        );
-        assert_eq!(
-            text_values(
-                extracted
-                    .subsequent
-                    .as_deref()
-                    .expect("subsequent variant should exist")
-            ),
-            vec!["A", "SHORT", "MID", "DATE", "Z"]
-        );
-        assert_eq!(
-            text_values(
-                extracted
-                    .ibid
-                    .as_deref()
-                    .expect("ibid variant should exist")
-            ),
-            vec!["A", "FULL", "MID", "IBID", "Z"]
-        );
-    }
-
-    #[test]
-    fn extract_position_templates_selects_ibid_per_choose() {
-        let ibid_only_choose = Choose {
-            if_branch: choose_branch(Some("ibid"), vec![literal_text("IBID-ONLY")]),
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("DEFAULT-A")]),
-        };
-        let ibid_with_locator_choose = Choose {
-            if_branch: choose_branch(
-                Some("ibid-with-locator"),
-                vec![literal_text("IBID-WITH-LOC")],
-            ),
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("DEFAULT-B")]),
-        };
-        let legacy_nodes = vec![
-            CslNode::Choose(ibid_only_choose),
-            literal_text("MID"),
-            CslNode::Choose(ibid_with_locator_choose),
-        ];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert_eq!(
-            text_values(
-                extracted
-                    .ibid
-                    .as_deref()
-                    .expect("ibid variant should exist")
-            ),
-            vec!["IBID-ONLY", "MID", "IBID-WITH-LOC"]
-        );
-    }
-
-    #[test]
-    fn extract_position_templates_preserves_nested_non_position_choose() {
-        let nested_choose = Choose {
-            if_branch: ChooseBranch {
-                match_mode: None,
-                type_: None,
-                variable: Some("title".to_string()),
-                is_numeric: None,
-                is_uncertain_date: None,
-                locator: None,
-                position: None,
-                children: vec![literal_text("LEGAL")],
-            },
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("GENERAL")]),
-        };
-        let choose = Choose {
-            if_branch: choose_branch(Some("subsequent"), vec![CslNode::Choose(nested_choose)]),
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-        let legacy_nodes = vec![CslNode::Choose(choose)];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert!(matches!(
-            extracted.subsequent.as_deref(),
-            Some([csln::CslnNode::Condition(_)])
-        ));
-    }
-
-    #[test]
-    fn extract_position_templates_rewrites_position_type_branches_into_conditionals() {
-        let choose = Choose {
-            if_branch: choose_branch_with_conditions(
-                None,
-                None,
-                Some("archive"),
-                None,
-                vec![literal_text("ARCHIVE")],
-            ),
-            else_if_branches: vec![
-                choose_branch_with_conditions(
-                    Some("first"),
-                    Some("interview"),
-                    None,
-                    None,
-                    vec![literal_text("INTERVIEW-FIRST")],
-                ),
-                choose_branch_with_conditions(
-                    Some("first"),
-                    Some("personal_communication"),
-                    None,
-                    None,
-                    vec![literal_text("PERSONAL-FIRST")],
-                ),
-            ],
-            else_branch: None,
-        };
-        let legacy_nodes = vec![CslNode::Choose(choose)];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert!(matches!(
-            extracted.first.as_deref(),
-            Some([csln::CslnNode::Condition(condition)])
-                if condition.if_variables.len() == 1
-                    && condition.else_if_branches.len() == 2
-                    && condition.else_if_branches.iter().all(|branch| branch.if_item_type.len() == 1)
-                    && condition.else_branch.is_none()
-        ));
-        assert_eq!(
-            text_values(
-                extracted
-                    .first
-                    .as_deref()
-                    .expect("first variant should exist")
-            ),
-            vec!["ARCHIVE", "INTERVIEW-FIRST", "PERSONAL-FIRST"]
-        );
-    }
-
-    #[test]
-    fn extract_position_templates_preserves_non_position_siblings_in_mixed_subsequent_variant() {
-        let choose = Choose {
-            if_branch: choose_branch_with_conditions(
-                None,
-                None,
-                Some("archive"),
-                None,
-                vec![literal_text("ARCHIVE")],
-            ),
-            else_if_branches: vec![
-                choose_branch_with_conditions(
-                    Some("subsequent"),
-                    Some("interview"),
-                    None,
-                    None,
-                    vec![literal_text("SHORT-INTERVIEW")],
-                ),
-                choose_branch_with_conditions(
-                    Some("subsequent"),
-                    Some("personal_communication"),
-                    None,
-                    None,
-                    vec![literal_text("SHORT-PERSONAL")],
-                ),
-            ],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-        let legacy_nodes = vec![group(vec![
-            literal_text("PREFIX"),
-            CslNode::Choose(choose),
-            literal_text("SUFFIX"),
-        ])];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert!(matches!(
-            extracted.subsequent.as_deref(),
-            Some([csln::CslnNode::Group(group)])
-                if matches!(group.children.as_slice(), [_, csln::CslnNode::Condition(_), _])
-        ));
-        assert_eq!(
-            text_values(
-                extracted
-                    .subsequent
-                    .as_deref()
-                    .expect("subsequent variant should exist")
-            ),
-            vec![
-                "PREFIX",
-                "ARCHIVE",
-                "SHORT-INTERVIEW",
-                "SHORT-PERSONAL",
-                "FULL",
-                "SUFFIX"
-            ]
-        );
-    }
-
-    #[test]
-    fn extract_position_templates_supports_position_locator_branches() {
-        let choose = Choose {
-            if_branch: choose_branch_with_conditions(
-                Some("first"),
-                None,
-                None,
-                Some("chapter"),
-                vec![literal_text("CHAPTER-FIRST")],
-            ),
-            else_if_branches: vec![choose_branch_with_conditions(
-                Some("subsequent"),
-                None,
-                None,
-                Some("page"),
-                vec![literal_text("PAGE-SUBSEQUENT")],
-            )],
-            else_branch: Some(vec![literal_text("FALLBACK")]),
-        };
-        let legacy_nodes = vec![CslNode::Choose(choose)];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(!extracted.unsupported_mixed_conditions);
-        assert_eq!(
-            text_values(
-                extracted
-                    .first
-                    .as_deref()
-                    .expect("first variant should exist")
-            ),
-            vec!["CHAPTER-FIRST", "FALLBACK"]
-        );
-        assert_eq!(
-            text_values(
-                extracted
-                    .subsequent
-                    .as_deref()
-                    .expect("subsequent variant should exist")
-            ),
-            vec!["PAGE-SUBSEQUENT", "FALLBACK"]
-        );
-    }
-
-    #[test]
-    fn extract_position_templates_marks_ambiguous_fallbacks_as_unsupported() {
-        let choose = Choose {
-            if_branch: choose_branch_with_conditions(
-                None,
-                None,
-                Some("title"),
-                None,
-                vec![literal_text("TITLE")],
-            ),
-            else_if_branches: vec![choose_branch(
-                Some("subsequent"),
-                vec![literal_text("SHORT")],
-            )],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-        let legacy_nodes = vec![CslNode::Choose(choose)];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(extracted.unsupported_mixed_conditions);
-        assert!(!extracted.has_overrides());
-        assert!(extracted.first.is_none());
-    }
-
-    #[test]
-    fn extract_position_templates_marks_duplicate_position_branches_as_unsupported() {
-        let choose = Choose {
-            if_branch: choose_branch(Some("subsequent"), vec![literal_text("SHORT")]),
-            else_if_branches: vec![choose_branch(
-                Some("subsequent"),
-                vec![literal_text("AGAIN")],
-            )],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-        let legacy_nodes = vec![CslNode::Choose(choose)];
-
-        let extracted = Upsampler::new().extract_citation_position_templates(&legacy_nodes);
-
-        assert!(extracted.unsupported_mixed_conditions);
-    }
-
-    #[test]
-    fn map_choose_prefers_else_branch_for_uncertain_dates() {
-        let choose = Choose {
-            if_branch: ChooseBranch {
-                match_mode: None,
-                type_: None,
-                variable: None,
-                is_numeric: None,
-                is_uncertain_date: Some("issued".to_string()),
-                locator: None,
-                position: None,
-                children: vec![literal_text("UNCERTAIN")],
-            },
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("CERTAIN")]),
-        };
-
-        let mapped = Upsampler::new()
-            .map_choose(&choose)
-            .expect("uncertain-date choose should map");
-
-        assert_eq!(text_values(std::slice::from_ref(&mapped)), vec!["CERTAIN"]);
-    }
-
-    #[test]
-    fn map_choose_drops_uncertain_date_markers_without_default_branch() {
-        let choose = Choose {
-            if_branch: ChooseBranch {
-                match_mode: None,
-                type_: None,
-                variable: None,
-                is_numeric: None,
-                is_uncertain_date: Some("issued".to_string()),
-                locator: None,
-                position: None,
-                children: vec![literal_text("circa")],
-            },
-            else_if_branches: vec![],
-            else_branch: None,
-        };
-
-        let mapped = Upsampler::new()
-            .map_choose(&choose)
-            .expect("uncertain-date choose should map");
-
-        match mapped {
-            csln::CslnNode::Group(group) => assert!(group.children.is_empty()),
-            other => panic!("expected empty group fallback, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn map_choose_prefers_first_citation_content_for_position_fallbacks() {
-        let choose = Choose {
-            if_branch: choose_branch(Some("subsequent"), vec![literal_text("SHORT")]),
-            else_if_branches: vec![],
-            else_branch: Some(vec![literal_text("FULL")]),
-        };
-
-        let mapped = Upsampler::new()
-            .map_choose(&choose)
-            .expect("position choose should map");
-
-        assert_eq!(text_values(std::slice::from_ref(&mapped)), vec!["FULL"]);
-    }
-
-    #[test]
-    fn map_choose_uses_negated_else_if_as_effective_fallback() {
-        let choose = Choose {
-            if_branch: choose_branch_with_conditions(
-                None,
-                Some("book"),
-                None,
-                None,
-                vec![literal_text("BOOK")],
-            ),
-            else_if_branches: vec![ChooseBranch {
-                match_mode: Some("none".to_string()),
-                type_: Some("book".to_string()),
-                variable: None,
-                is_numeric: None,
-                is_uncertain_date: None,
-                locator: None,
-                position: None,
-                children: vec![literal_text("NOT-BOOK")],
-            }],
-            else_branch: None,
-        };
-
-        let mapped = Upsampler::new()
-            .map_choose(&choose)
-            .expect("choose should map");
-
-        match mapped {
-            csln::CslnNode::Condition(condition) => {
-                assert!(condition.else_if_branches.is_empty());
-                assert_eq!(text_values(&condition.then_branch), vec!["BOOK"]);
-                assert_eq!(
-                    text_values(
-                        condition
-                            .else_branch
-                            .as_deref()
-                            .expect("negated else-if should become else branch")
-                    ),
-                    vec!["NOT-BOOK"]
-                );
-            }
-            other => panic!("expected condition node, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn map_choose_preserves_existing_else_over_negated_if_fallback() {
-        let choose = Choose {
-            if_branch: ChooseBranch {
-                match_mode: Some("none".to_string()),
-                type_: Some("book".to_string()),
-                variable: None,
-                is_numeric: None,
-                is_uncertain_date: None,
-                locator: None,
-                position: None,
-                children: vec![literal_text("NOT-BOOK")],
-            },
-            else_if_branches: vec![choose_branch_with_conditions(
-                None,
-                Some("article-journal"),
-                None,
-                None,
-                vec![literal_text("ARTICLE")],
-            )],
-            else_branch: Some(vec![literal_text("EXISTING-ELSE")]),
-        };
-
-        let mapped = Upsampler::new()
-            .map_choose(&choose)
-            .expect("choose should map");
-
-        match mapped {
-            csln::CslnNode::Condition(condition) => {
-                assert!(condition.then_branch.is_empty());
-                assert_eq!(
-                    text_values(
-                        condition
-                            .else_branch
-                            .as_deref()
-                            .expect("existing else should remain in place")
-                    ),
-                    vec!["EXISTING-ELSE"]
-                );
-                let mut all_text = text_values(
-                    condition
-                        .else_branch
-                        .as_deref()
-                        .expect("existing else should remain in place"),
-                );
-                all_text.extend(text_values(&condition.then_branch));
-                for branch in &condition.else_if_branches {
-                    all_text.extend(text_values(&branch.children));
-                }
-                assert!(!all_text.contains(&"NOT-BOOK"));
-                assert!(all_text.contains(&"ARTICLE"));
-            }
-            other => panic!("expected condition node, got {other:?}"),
-        }
+        citum_nodes
     }
 }
